Auto-restart if node space is too low.
Ugh - finding the ESP module automatically resetting itself; previous debug
stuff showed it was an exception, not power related, which made me suspect
memory exhaustion.

Adding reporting of heap space to status output and via syslog after each
request shows that, yes, it does leak, and eventually gets too low.

After a chat in irc://irc.freenode.net/esp8266 it seems a common problem with no
real solution - that the heap fragments, and eventually there's no space.

So, hackish solution - check the heap space, and automatically reboot if we're
running out and the output pin is low (don't do it if it is high, as we
don't want to turn off the relay).

TODO: if reverse_logic is in effect, this may not work so well - if we've been
told to put the thing on, we'll pull the pin low as the logic is negated - this
new code may consider that an opportune time for a restart, at which point upon
restart, we'd pull it high.  Needs more thought.

The problem was made much worse for me because I added status polling to my
internal management page, so each client with the page open would poll the ESP
every 20 seconds for the status - each request leaking a little more heap space.
So, I'll stop that, that should give much better uptime.  Explains why I only
started having this trouble recently!

Still interested if there are any decent solutions to avoid heap fragmentation
issues, though.